// task-manager/src/app/core/models/tasks.model.ts
export type TaskStatus = "pending" | "completed";

export interface Task {
    _id: string;
    title: string;
    description: string;
    status: TaskStatus;
    createdAt: string;
    updatedAt: string;
}

export interface CreateTaskDto { title: string; description?: string; status?: TaskStatus; }
export interface UpdateTaskDto { title?: string; description?: string; status?: TaskStatus; }